+++ vector/src/main/java/im/vector/activity/VectorRoomActivity.java
@@ -1258,8 +1258,7 @@ public class VectorRoomActivity extends MXCActionBarActivity implements
         // to have notifications for this room
         ViewedRoomTracker.getInstance().setViewedRoomId(null);
         ViewedRoomTracker.getInstance().setMatrixId(null);
-        //mEditText.initAutoCompletion(mSession, null);
-        //mEditText.initAutoCompletionCommandLine(mSession, null);
+        mEditText.initAutoCompletion(mSession, null);
     }
 
     @Override
@@ -1395,8 +1394,7 @@ public class VectorRoomActivity extends MXCActionBarActivity implements
         displayE2eRoomAlert();
 
         // init the auto-completion list from the room members
-        //mEditText.initAutoCompletion(mSession, (null != mRoom) ? mRoom.getRoomId() : null);
-        //mEditText.initAutoCompletionCommandLine(mSession, (null != mRoom) ? mRoom.getRoomId() : null);
+        mEditText.initAutoCompletion(mSession, (null != mRoom) ? mRoom.getRoomId() : null);
 
         if (mReadMarkerManager != null) {
             mReadMarkerManager.onResume();