+++ vector/src/main/java/im/vector/util/SlashCommandsParser.java
@@ -17,6 +17,7 @@
 
 package im.vector.util;
 
+import android.content.res.Resources;
 import android.support.v7.app.AlertDialog;
 import android.text.TextUtils;
 import android.widget.Toast;
@@ -30,6 +31,7 @@ import org.matrix.androidsdk.rest.callback.SimpleApiCallback;
 import org.matrix.androidsdk.rest.model.MatrixError;
 import org.matrix.androidsdk.util.Log;
 
+import java.util.ArrayList;
 import java.util.Collection;
 import java.util.HashMap;
 import java.util.List;
@@ -38,6 +40,7 @@ import java.util.Map;
 import im.vector.R;
 import im.vector.VectorApp;
 import im.vector.activity.CommonActivityUtils;
+import im.vector.activity.RiotAppCompatActivity;
 import im.vector.activity.VectorRoomActivity;
 import im.vector.widgets.WidgetsManager;
 
@@ -48,36 +51,58 @@ public class SlashCommandsParser {
     // defines the command line operations
     // the user can write theses messages to perform some room events
     public static final String CMD_EMOTE = "/me";
+    public static final String PARAM_EMOTE ="<message>";
+    public static final String DESC_EMOTE ="Displays action";
 
     // <user-id> [reason]
     public static final String CMD_BAN_USER = "/ban";
+    public static final String PARAM_BAN_USER ="<user-id>";
+    public static final String DESC_BAN_USE ="Bans user with given id";
 
     // <user-id>'
     public static final String CMD_UNBAN_USER = "/unban";
+    public static final String PARAM_UNBAN_USER = "<user-id>";
+    public static final String DESC_UNBAN_USER = "Unbans user with given id";
 
     // <user-id> [<power-level>]
     public static final String CMD_SET_USER_POWER_LEVEL = "/op";
+    public static final String PARAM_SET_USER_POWER_LEVEL = "<user-id> [<power-level>]";
+    public static final String DESC_SET_USER_POWER_LEVEL = "Define the power level of a user";
 
     // <user-id>
     public static final String CMD_RESET_USER_POWER_LEVEL = "/deop";
+    public static final String PARAM_RESET_USER_POWER_LEVEL = "<user-id>";
+    public static final String DESC_RESET_USER_POWER_LEVEL = "Deops user with given id";
 
     // <user-id>
     public static final String CMD_INVITE = "/invite";
+    public static final String PARAM_INVITE = "<user-id>";
+    public static final String DESC_INVITE = "Invites user with given id to current room";
 
     // <room-alias>
     public static final String CMD_JOIN_ROOM = "/join";
+    public static final String PARAM_JOIN_ROOM = "<room-alias>";
+    public static final String DESC_JOIN_ROOM = "Joins room with given alias";
 
     // <room-alias>
     public static final String CMD_PART = "/part";
+    public static final String PARAM_PART = "<room-alias>";
+    public static final String DESC_PART = "Leave room";
 
     // <topic>
     public static final String CMD_TOPIC = "/topic";
+    public static final String PARAM_TOPIC = "<topic>";
+    public static final String DESC_TOPIC = "Set the room topic";
 
     // <user-id> [reason]
     public static final String CMD_KICK_USER = "/kick";
+    public static final String PARAM_KICK_USER = "<user-id>";
+    public static final String DESC_KICK_USER = "Kicks user with given id";
 
     // <display-name>
     public static final String CMD_CHANGE_DISPLAY_NAME = "/nick";
+    public static final String PARAM_CHANGE_DISPLAY_NAME = "<display-name>";
+    public static final String DESC_CHANGE_DISPLAY_NAME = "Changes your display nickname";
 
     // <query>
     private static final String CMD_DDG = "/ddg";
@@ -96,16 +121,99 @@ public class SlashCommandsParser {
 
     // on / off
     public static final String CMD_MARKDOWN = "/markdown";
+    public static final String PARAM_MARKDOWN = "";
+    public static final String DESC_MARKDOWN = "On/Off markdown";
 
     // clear scalar token (waiting for correct 403 management)
     public static final String CMD_CLEAR_SCALAR_TOKEN = "/clear_scalar_token";
+    public static final String PARAM_CLEAR_SCALAR_TOKEN = "";
+    public static final String DESC_CLEAR_SCALAR_TOKEN = "To fix Matrix Apps management";
 
-    public String slashCommand;
     public String parameter;
     public String description;
-    public List<SlashCommandsParser> slashCommandList;
+    public static List<String> mSlashCommandList;
+    public static HashMap<String, String> mSlashCommandParamMap;
+    public static HashMap<String, String> mSlashCommandDescMap;
 
-    //TODO return a list of command lines
+    public static List<String> getSlashCommandList() {
+        createSlashCommandList();
+        setSlashCommandParameter();
+        setSlashCommandDescription();
+        return mSlashCommandList;
+    }
+
+    public static String getSlashCommandParam(String command) {
+
+        if (mSlashCommandParamMap.containsKey(command)) {
+            return mSlashCommandParamMap.get(command);
+        }
+
+        return null;
+    }
+
+    public static String getSlashCommandDescription(String command) {
+
+        if (mSlashCommandDescMap.containsKey(command)) {
+            return mSlashCommandDescMap.get(command);
+        }
+
+        return null;
+    }
+
+    private static void createSlashCommandList() {
+        mSlashCommandList = new ArrayList<String>();
+
+        mSlashCommandList.add(CMD_EMOTE);
+        mSlashCommandList.add(CMD_CHANGE_DISPLAY_NAME);
+        mSlashCommandList.add(CMD_TOPIC);
+        mSlashCommandList.add(CMD_INVITE);
+        mSlashCommandList.add(CMD_JOIN_ROOM);
+        mSlashCommandList.add(CMD_PART);
+        mSlashCommandList.add(CMD_SET_USER_POWER_LEVEL);
+        mSlashCommandList.add(CMD_RESET_USER_POWER_LEVEL);
+        mSlashCommandList.add(CMD_KICK_USER);
+        mSlashCommandList.add(CMD_BAN_USER);
+        mSlashCommandList.add(CMD_UNBAN_USER);
+        mSlashCommandList.add(CMD_MARKDOWN);
+        mSlashCommandList.add(CMD_CLEAR_SCALAR_TOKEN);
+    }
+
+    private static void setSlashCommandParameter(){
+        mSlashCommandParamMap = new HashMap<>();
+
+        mSlashCommandParamMap.put(CMD_EMOTE, PARAM_EMOTE);
+        mSlashCommandParamMap.put(CMD_CHANGE_DISPLAY_NAME, PARAM_CHANGE_DISPLAY_NAME);
+        mSlashCommandParamMap.put(CMD_TOPIC, PARAM_TOPIC);
+        mSlashCommandParamMap.put(CMD_INVITE, PARAM_INVITE);
+        mSlashCommandParamMap.put(CMD_JOIN_ROOM, PARAM_JOIN_ROOM);
+        mSlashCommandParamMap.put(CMD_PART, PARAM_PART);
+        mSlashCommandParamMap.put(CMD_SET_USER_POWER_LEVEL, PARAM_SET_USER_POWER_LEVEL);
+        mSlashCommandParamMap.put(CMD_RESET_USER_POWER_LEVEL, PARAM_RESET_USER_POWER_LEVEL);
+        mSlashCommandParamMap.put(CMD_KICK_USER, PARAM_KICK_USER);
+        mSlashCommandParamMap.put(CMD_BAN_USER, PARAM_BAN_USER);
+        mSlashCommandParamMap.put(CMD_UNBAN_USER, PARAM_UNBAN_USER);
+        mSlashCommandParamMap.put(CMD_MARKDOWN, PARAM_MARKDOWN);
+        mSlashCommandParamMap.put(CMD_CLEAR_SCALAR_TOKEN, PARAM_CLEAR_SCALAR_TOKEN);
+
+    }
+
+    private static void setSlashCommandDescription() {
+        mSlashCommandDescMap = new HashMap<>();
+
+        mSlashCommandDescMap.put(CMD_EMOTE, DESC_EMOTE);
+        mSlashCommandDescMap.put(CMD_CHANGE_DISPLAY_NAME, DESC_CHANGE_DISPLAY_NAME);
+        mSlashCommandDescMap.put(CMD_TOPIC, DESC_TOPIC);
+        mSlashCommandDescMap.put(CMD_INVITE, DESC_INVITE);
+        mSlashCommandDescMap.put(CMD_JOIN_ROOM, DESC_JOIN_ROOM);
+        mSlashCommandDescMap.put(CMD_PART, DESC_PART);
+        mSlashCommandDescMap.put(CMD_SET_USER_POWER_LEVEL, DESC_SET_USER_POWER_LEVEL);
+        mSlashCommandDescMap.put(CMD_RESET_USER_POWER_LEVEL, DESC_RESET_USER_POWER_LEVEL);
+        mSlashCommandDescMap.put(CMD_KICK_USER, DESC_KICK_USER);
+        mSlashCommandDescMap.put(CMD_BAN_USER, DESC_BAN_USE);
+        mSlashCommandDescMap.put(CMD_UNBAN_USER, DESC_UNBAN_USER);
+        mSlashCommandDescMap.put(CMD_MARKDOWN, DESC_MARKDOWN);
+        mSlashCommandDescMap.put(CMD_CLEAR_SCALAR_TOKEN, DESC_CLEAR_SCALAR_TOKEN);
+    }
 
     /**
      * check if the text message is an IRC command.